Refactor breakpoints actions (#3289)
